fix: refactor after review
