+++ lib/features/departments/department_detail_view/data/repository/department_details_repository.dart
@@ -5,22 +5,23 @@ import "../../../../../api_base_rest/cache/cache.dart";
 import "../../../../../config/env.dart";
 import "../../../../../config/ttl_config.dart";
 import "../../presentation/department_detail_view.dart";
-import "../models/detailed_department.dart";
+import "../models/department_details.dart";
 
 part "department_details_repository.g.dart";
 
 @riverpod
-Future<DetailedDepartment> departmentDetailsRepository2(Ref ref, int id) async {
-  final url = "${Env.mainRestApiUrl}/departments/$id?fieldOfStudy=true&departmentLink=true";
+Future<DepartmentDetails> departmentDetailsRepository(Ref ref, int id) async {
+  final endpoint = "/departments/$id?fieldOfStudy=true&departmentLink=true";
+  final url = "${Env.mainRestApiUrl}$endpoint";
 
   final response = await ref.getAndCacheData(
     url,
     TtlStrategy.get(TtlKey.departmentDetailsRepository).inDays,
-    DetailedDepartment.fromJson,
+    DepartmentDetailsResponse.fromJson,
     extraValidityCheck: (_) => true,
     localizedOfflineMessage: DepartmentDetailView.localizedOfflineMessage,
     onRetry: ref.invalidateSelf,
   );
 
-  return response;
+  return response.data;
 }
