+++ lib/features/departments/department_detail_view/data/utils/department_detail_config.dart
@@ -1,0 +1,4 @@
+abstract class DepartmentDetailConfig {
+  static const longCycleStudiesSemesterCount = 12;
+  static const separateAddressPattern = r"^(.*)\s((ul\.|wyb\.|pl\.).*)";
+}
