+++ lib/features/departments/department_detail_view/presentation/widgets/field_of_study_expansion_tile.dart
@@ -3,7 +3,7 @@ import "package:separate/separate.dart";
 
 import "../../../../../theme/app_theme.dart";
 import "../../../../../widgets/my_expansion_tile.dart";
-import "../../data/models/field_of_study.dart";
+import "../../data/models/department_details.dart";
 import "field_of_study_tile.dart";
 
 class FieldOfStudyExpansionTile extends StatelessWidget {
