+++ lib/features/departments/departments_view/data/models/department.dart
@@ -2,10 +2,15 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter/foundation.dart";
 import "package:freezed_annotation/freezed_annotation.dart";
 
-import "../../../../../api_base_rest/utils/json_parse.dart";
-
 part "department.freezed.dart";
 part "department.g.dart";
+
+@freezed
+abstract class DepartmentsResponse with _$DepartmentsResponse {
+  const factory DepartmentsResponse({required IList<Department> data}) = _DepartmentsResponse;
+
+  factory DepartmentsResponse.fromJson(Map<String, dynamic> json) => _$DepartmentsResponseFromJson(json);
+}
 
 @freezed
 abstract class Department with _$Department {
@@ -18,14 +23,6 @@ abstract class Department with _$Department {
     required String gradientStart,
     required String gradientStop,
   }) = _Department;
-
-  static IList<Department> fromJsonList(Map<String, dynamic> json) {
-    return json
-        .parseJsonToList()
-        .map((departmentJson) => Department.fromJson(departmentJson as Map<String, dynamic>))
-        .toList()
-        .lock;
-  }
 
   factory Department.fromJson(Map<String, dynamic> json) => _$DepartmentFromJson(json);
 }
